WIP: expose `toggleLink` as public method
